conditionally compile go1.7 to preserve < 1.7 behavior
