+++ helper_15.go
@@ -1,0 +1,40 @@
+// +build !go1.7
+
+/********************************
+*** Multiplexer for Go        ***
+*** Bone is under MIT license ***
+*** Code by CodingFerret      ***
+*** github.com/go-zoo         ***
+*********************************/
+
+package bone
+
+import (
+	"net/http"
+	"sync"
+)
+
+var globalVars = struct {
+	sync.RWMutex
+	v map[*http.Request]map[string]string
+}{v: make(map[*http.Request]map[string]string)}
+
+// GetAllValues return the req PARAMs
+func GetAllValues(req *http.Request) map[string]string {
+	globalVars.RLock()
+	values := globalVars.v[req]
+	globalVars.RUnlock()
+	return values
+}
+
+// serveMatchedRequest is an extension point for Route which allows us to conditionally compile for
+// go1.7 and <go1.7
+func (r *Route) serveMatchedRequest(rw http.ResponseWriter, req *http.Request, vars map[string]string) {
+	globalVars.Lock()
+	globalVars.v[req] = vars
+	globalVars.Unlock()
+	r.Handler.ServeHTTP(rw, req)
+	globalVars.Lock()
+	delete(globalVars.v, req)
+	globalVars.Unlock()
+}
